Add memo na rota padrão, muitos requests nos endpoints de listagem de pendências
