+++ src/SME.SGP.WebClient/src/paginas/Principal/principal.js
@@ -1,61 +1,12 @@
-import React, { useCallback, useEffect, useState } from 'react';
-import { useSelector } from 'react-redux';
+import React from 'react';
 import CardsDashboard from '~/componentes-sgp/cardsDashboard/cardsDashboard';
-import Alert from '../../componentes/alert';
+import AlertasFiltroPrincipal from './alertasFiltroPrincipal';
 import PendenciasGerais from './Pendencias/pendenciasGerais';
 
 const Principal = () => {
-  const [turmaSelecionada, setTurmaSelecionada] = useState(false);
-
-  const usuario = useSelector(state => state.usuario);
-  const perfil = useSelector(state => state.perfil.perfilSelecionado);
-  const modalidades = useSelector(state => state.filtro.modalidades);
-  const anosLetivo = useSelector(state => state.filtro.anosLetivos);
-
-  const validarFiltro = useCallback(() => {
-    if (!usuario.turmaSelecionada) {
-      setTurmaSelecionada(false);
-      return;
-    }
-
-    const temTurma = !!usuario.turmaSelecionada.turma;
-
-    setTurmaSelecionada(temTurma);
-  }, [usuario.turmaSelecionada]);
-
-  useEffect(() => {
-    validarFiltro();
-  }, [usuario, validarFiltro]);
-
   return (
     <div>
-      {modalidades &&
-      !modalidades.length &&
-      anosLetivo.length === 1 &&
-      !usuario.ehPerfilProfessor &&
-      perfil &&
-      perfil.nomePerfil === 'Supervisor' ? (
-        <Alert
-          alerta={{
-            tipo: 'warning',
-            id: 'AlertaPrincipal',
-            mensagem: `Não foi possível obter as escolas atribuídas ao supervisor ${usuario.rf}`,
-          }}
-        />
-      ) : (
-        <></>
-      )}
-      {!turmaSelecionada ? (
-        <Alert
-          alerta={{
-            tipo: 'warning',
-            id: 'AlertaPrincipal',
-            mensagem: 'Você precisa escolher uma turma',
-          }}
-        />
-      ) : (
-        <></>
-      )}
+      <AlertasFiltroPrincipal />
       <CardsDashboard />
       <PendenciasGerais />
     </div>
